Importer split mesh blender into own file
